watch branches of the main repo that the config is read from
The repo that the config exists on can have multiple branches. Each
branch will be watched for new SHAs. When checking out the repo from
the config, it will attempt to checkout the given branch. If that
fails, it will fallback to the default branch to run the script
against.