Adds Parameter field to Task configuration
Parameters are used to export environment variables within the script.
The environment variables are forwarded from the pushed app. This is
useful for secrets that should not be commited in a repo. They must be
formatted as:

tasks:
- repo_path: https://github.com/apoydence/triple-c
  parameters:
    SOME_SECRET: ((CF_PUSH_SECRET))
    OTHER: known_value
  command: echo "$(SOME_SECRET) $(OTHER)"

In the above example, ((CF_PUSH_SECRET)) is replaced
by whatever the app has exported for CF_PUSH_SECRET.