Allow tasks to share data.
Adds 'output' and 'input' fields to tasks.  If a task has an output, it
will have a directory available to dump data into. At the end of a
successful task, the directory will be tar'd up and POSTed to the
server. The next task will grab and untar the directory. It will then
move the output to its configured 'input' directory.